requirements.txt
discord.py
disnake
mysql-connector-python